% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt.R
\name{filt}
\alias{filt}
\title{Delineating movement trips for central-place foraging species}
\usage{
filt(
  pathF = ...,
  pathM = ...,
  metname = NULL,
  nbn = c("NULL"),
  gpst = NULL,
  ddep = NULL,
  drecap = NULL,
  colony = NULL,
  year = NULL,
  ring = NULL,
  tdep = NULL,
  trecap = NULL,
  timezone = NULL,
  Clong = NULL,
  Clat = NULL,
  speedTresh = NULL,
  FIX = NULL,
  FixInt = NULL,
  BuffColony = NULL,
  MinTripDur = NULL,
  Complete = FALSE,
  Interpolate = FALSE,
  filtNA = 1,
  metINFO = c(NULL),
  splt = TRUE
)
}
\arguments{
\item{pathF}{path leading to your tracks}

\item{pathM}{path leading to your metadata file}

\item{metname}{a character string corresponding to the metadata file's name}

\item{nbn}{character. Vector of "names" matching your raw GPS files (i.e. which are linked by the pathF argument). See details for more details}

\item{gpst}{a character string corresponding to the metadata's column name containing the GPS type}

\item{ddep}{a character string corresponding to the metadata's column name containing the date of departure}

\item{drecap}{a character string corresponding to the metadata's column name containing the date of retrieval}

\item{colony}{a character string corresponding to the metadata's column name containing individual colony attribution}

\item{year}{a character string corresponding to the metadata's column name containing the year of sampling}

\item{ring}{a character string corresponding to the metadata's column name containing the individual ring number}

\item{tdep}{a character string corresponding to the metadata's column name containing the time of departure}

\item{trecap}{a character string corresponding to the metadata's column name containing the time of retrieval}

\item{timezone}{time zone in which data are recorded}

\item{speedTresh}{numeric treshold used as speed cutoff for speed filtering (km/h)}

\item{FIX}{a character string corresponding to the metadata's column name containing the numeric interval separating two
locations in the GPS track}

\item{FixInt}{numeric interval that should separate two locations in GPS tracks for interpolation (minutes; e.g. 2, 10)}

\item{BuffColony}{numeric value indicating the buffer radius length around the colony (km)}

\item{MinTripDur}{numeric value used for filtering trip by minimum trip duration (minutes)}

\item{Complete}{logical; if TRUE, only complete trip will be kept}

\item{Interpolate}{logical; if TRUE, tracks are interpolated}

\item{filtNA}{numeric value which correspond of the filtering treshold for number of NAs generated along the
interpolation. Between 0 and 1, whereas 1 correspond to no filtration at all (filtering only trips that are 100\% full of NAs)}

\item{metINFO}{character string(s). Used as a vector of name for the additional information you want to add to the final
data frame for facilitating the subsequent analyses. However, names should fit the column name occuring in the metadata file (i.e. \emph{pathM})}

\item{splt}{logical; if TRUE, the final trips are ordered as a list instead of an unique data frame. Default FALSE}
}
\value{
the function return either a data frame (splt = FALSE) or a list (splt = TRUE) containing each bird trip as component
Metrics for each trip are also computed, as;
\itemize{
\item track_ID: is the unique track ID, should be named following each GPS raw file
\item nbNA: indicating if the relative point is precise or if is was NA after the interpolation process
\item ColonyDist: is the distance (km) separating the relative point and the colony
\item PointDist: is the distance (km) separating two consecutive points (zi - zi-1)
\item tripID: trip number in the trackID group
\item birdTrip: unique bird trip identifyer
\item TripLength: Total time length of the trip (min)
\item TripDist: Cumulative distance flew by the bird for the specific trip
\item Speed: Speed between t-1 and t in m/s
\item nPoints: number of points composing the trip
\item maxDist: maximum distance that the bird went in its trip, relative to the colony
\item propNA: proportion of NAs in the trip. Convenient for filtering for reliability of the statistics
\item BegPoint: colony or trip. Specify if the first point was at the colony or outside (i.e. trip). Convenient for filtering for completness
\item EndPoint: colony or trip. Specify if the last point was at the colony or outside (i.e. trip). Convenient for filtering for completness
}
}
\description{
This function was developped to conveniently delineate trips from raw GPS data for
central-place foraging species. It also includes various "filtering" options to segment trips.
}
\details{
Raw GPS data (located via \emph{pathF}) should be as .csv format. This version only include 4 types of GPS format 1) Catlog, 2)
IGotU, 3) PathTracks and 4) Ecotone. If your file doesn't have any specific format, the \emph{gpst} could be specified as IGotU (one line header,
and the raw data in the second lines. However, one should make sure that the \emph{Latitude}, \emph{Longitude}, \emph{Date} and \emph{Time} are specified.
\emph{Date} could be supplied as character, as YYYY-MM-DD or MM/DD/YYYY and \emph{Time} also as character as HH:MM:SS). Raw GPS data are expected to be in lat long coordinate system
Distance between points are calculated via the Great Circle distance, computed via \emph{sp} package
}
\examples{
\dontrun{
pathF <- c("C:/Users/User/files/allspecies/")
pathM <- c("C:/Users/User/files/metadata/")
metname <- c("meta.csv")
timezone <- c("GMT")

f <- filt(pathF, pathM, metname, nbn = c("year", "colony", "ring", "recapture"),
timezone, speedTresh = 90, gpst = "GPSType",
ddep = "deployment", drecap = "recapture", colony = "colony", year = "year",
ring = "ring", FIX = "FIX", tdep = "utc_deployment", trecap = "utc_retrieval",
Clong = "Clongitude", Clat = "Clatitude", BuffColony = 0.5, MinTripDur = 30, Complete = T, FixInt = 2,
Interpolate = T, filtNA = 0.9, metINFO = c("ring", "year", "species", "colony"), splt = F)
}
}
\references{
\itemize{
\item Freitas, C., Lydersen, C., Ims, R.A., Fedak, M.A. and Kovacs, K.M. (2008) A simple new algorithm to filter marine mammal Argos locations Marine Mammal Science 24:315-325.
\item McConnell, B.J., Chambers, C. and Fedak, M.A. (1992) Foraging ecology of southern elephant seals in relation to the bathymetry and productivity of the Southern Ocean. Antarctic Science 4:393-398.
}
}
\keyword{central-place}
\keyword{colony}
\keyword{cpf}
\keyword{delineating}
\keyword{forager}
\keyword{movement}
\keyword{trips}
