% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensTime.R
\name{sensTime}
\alias{sensTime}
\title{Sensitivity of minimum trip duration on trip delineation}
\usage{
sensTime(
  pathF = ...,
  pathM = ...,
  nbn = c("NULL"),
  iter = 50,
  metname = NULL,
  param = NULL,
  gpst = NULL,
  FIX = NULL,
  ddep = NULL,
  drecap = NULL,
  colony = NULL,
  year = NULL,
  ring = NULL,
  tdep = NULL,
  trecap = NULL,
  timezone = NULL,
  Clong = NULL,
  Clat = NULL,
  speedTresh = NULL,
  BuffColony = NULL,
  FixInt = NULL,
  Interpolate = FALSE,
  sorc = 0.05
)
}
\arguments{
\item{pathF}{path leading to your tracks}

\item{pathM}{path leading to your metadata file}

\item{nbn}{character. Vector of "names" matching your raw GPS files (i.e. which are linked by the pathF argument). See details for more details}

\item{iter}{the number of bootstrapping iterations (default = 50)}

\item{metname}{a character string corresponding to the metadata file's name}

\item{param}{vector of time treshold to be tested. These correspond to the different scenarios to be tested. Time is in hours format}

\item{gpst}{a character string corresponding to the metadata's column name containing the GPS type}

\item{FIX}{a character string corresponding to the metadata's column name containing the numeric interval separating two
locations in the GPS track}

\item{ddep}{a character string corresponding to the metadata's column name containing the date of departure}

\item{drecap}{a character string corresponding to the metadata's column name containing the date of retrieval}

\item{colony}{a character string corresponding to the metadata's column name containing individual colony attribution}

\item{year}{a character string corresponding to the metadata's column name containing the year of sampling}

\item{ring}{a character string corresponding to the metadata's column name containing the individual ring number}

\item{tdep}{a character string corresponding to the metadata's column name containing the time of departure}

\item{trecap}{a character string corresponding to the metadata's column name containing the time of retrieval}

\item{timezone}{time zone in which data are recorded}

\item{speedTresh}{numeric treshold used as speed cutoff for speed filtering (km/h)}

\item{BuffColony}{numeric value indicating the buffer radius length around the colony (km)}

\item{FixInt}{numeric interval that should separate two locations in GPS tracks for interpolation (minutes; e.g. 2, 10)}

\item{Interpolate}{logical; if TRUE, tracks are interpolated}

\item{sorc}{numeric, proportion (0:1) treshold used for the second-order-rate of change}
}
\value{
a list of object containing a data frame witht the main results of the sensitivity analysis and
the associated figures. Figures are also showing the "second-order-rate of change" for each scenario. For the
moment, the SORC is fixed to the 5\% absolute variaiton of the total magnitude of the response (i.e. max(average(nTrip))).
This is seen as a measure of "stability " of the response \emph{t} among scenario \emph{t-1} and \emph{t+1}.
}
\description{
This function was developped to give basic visualization tools to facilitate the selection
of the minimum trip length and the effect of this variation on the number of trip generated
}
\details{
Raw GPS data (located via \emph{pathF}) should be as .csv format. This version only include 4 types of GPS format 1) Catlog, 2)
IGotU, 3) PathTracks and 4) Ecotone. If your file doesn't have any specific format, the \emph{gpst} could be specified as IGotU (one line header,
and the raw data in the second lines. However, one should make sure that the \emph{Latitude}, \emph{Longitude}, \emph{Date} and \emph{Time} are specified.
\emph{Date} could be supplied as character, as YYYY-MM-DD or MM/DD/YYYY and \emph{Time} also as character as HH:MM:SS). Raw GPS data are expected to be in lat long coordinate system
Distance between points are calculated via the Great Circle distance, computed via \emph{sp} package
}
\examples{
\dontrun{
pathF <- c("C:/Users/philip/OneDrive/NP_Kontrakt/StromTracks/files/lomvi/")
pathM <- c("C:/Users/philip/OneDrive/NP_Kontrakt/StromTracks/metadata/")
metname <- c("meta_bjorn.csv")
timezone <- c("GMT")
param <- seq(0, 100, by = 10)

st <- sensTime(pathF, pathM, metname, nbn = c("year", "colony", "ring", "recapture"),
       timezone, iter = 300, param = param, speedTresh = 90, gpst = "GPSType", ddep = "deployment",
       drecap = "recapture", colony = "colony", year = "year", ring = "ring", FIX = "FIX",
       tdep = "utc_deployment", trecap = "utc_retrieval",
       Clong = "Clongitude", Clat = "Clatitude", BuffColony = 0.15, FixInt = 2, Interpolate = T, sorc = 0.01)
}
}
\references{
\itemize{
\item Freitas, C., Lydersen, C., Ims, R.A., Fedak, M.A. and Kovacs, K.M. (2008) A simple new algorithm to filter marine mammal Argos locations Marine Mammal Science 24:315-325.
\item McConnell, B.J., Chambers, C. and Fedak, M.A. (1992) Foraging ecology of southern elephant seals in relation to the bathymetry and productivity of the Southern Ocean. Antarctic Science 4:393-398.
}
}
\keyword{central-place-forager}
\keyword{colony}
\keyword{cpf}
\keyword{delineating}
\keyword{movement}
\keyword{movement-ecology}
\keyword{trips}
